+++ example/lib/core/widgets/setting_builder.dart
@@ -1,7 +1,7 @@
 import 'package:example/core/core.dart';
 import 'package:flutter/material.dart';
 
-enum SettingBuilder { debug, example2, alignment }
+enum SettingBuilder { debug, example2, alignment, scale }
 
 extension SettingBuilderExt on SettingBuilder {
   String get label {
@@ -12,6 +12,8 @@ extension SettingBuilderExt on SettingBuilder {
         return 'Visible example 2';
       case SettingBuilder.alignment:
         return 'Alignment';
+      case SettingBuilder.scale:
+        return 'scale';
     }
   }
 
@@ -24,19 +26,27 @@ extension SettingBuilderExt on SettingBuilder {
         return CustomListAlignment(
           label: label,
           value: value.$1,
-          onChanged: (e) => onChanged.call((e, value.$2, value.$3)),
+          onChanged: (e) => onChanged.call(value.copyWith(alignment: e)),
         );
       case SettingBuilder.debug:
         return CustomSwitchListTile(
           label: label,
-          value: value.$2.debug!,
-          onChanged: (e) => onChanged.call((value.$1, value.$2.copyWith(debug: e), value.$3)),
+          value: value.$2.debug,
+          onChanged: (e) => onChanged.call(value.copyWith(config: value.$2.copyWith(debug: e))),
         );
       case SettingBuilder.example2:
         return CustomSwitchListTile(
           label: label,
           value: value.$3,
-          onChanged: (e) => onChanged.call((value.$1, value.$2, e)),
+          onChanged: (e) => onChanged.call(value.copyWith(visible: e)),
+        );
+      case SettingBuilder.scale:
+        return CustomSlider(
+          label: label,
+          value: value.$4,
+          onChanged: (e) => onChanged.call(
+            value.copyWith(config: value.$2.adjustedScale(e), scale: e),
+          ),
         );
     }
   }